# switch packet packer to math/rand/v2